# implement numba colormap lookup