https://github.com/pyqtgraph/pyqtgraph/pull/1693#issuecomment-836241886 shows codepaths involving large luts ("uint16lut") are relatively slower. However, they are the codepaths that are taken by VideoSpeedTest.py and ImageView when using "float" and "uint16" dtypes.

Implementing the colormap lookup using Numba gives a decent speedup on Windows. YMMV on other platforms.

|      4096x4096 numba          |    master     |  this PR   |
|              -------------           | ------------- | ---------- |
| float_levels_uint16lut         |    69.9±0ms   |  28.4±0ms  |
| uint16_levels_uint16lut        |    50.7±0ms   |  19.0±0ms  |
| uint16_nolevels_uint16lut      |    50.6±0ms   |  19.0±0ms  |

It is surprising that a numba-ified version of ```np.take()``` should yield a 2.5x speedup.